Add gama-local tests to CMake build (by Kristian Evers)
This adds the ability to run unit tests for gama-local on CMake based
builds. The setup is kept as simple as possible and kept within
CMakeLists.txt files to avoid to much clutter in the test directories.

Running tests is simple:

cmake /path/to/repo
make
ctest

Autotools setup has been extended so that the newly add CMakeLists.txt
files are added to the source distribution when packaging a new release.

So far only a few tests has been added to the CMake test-suite but adding
more should be too big of a problem.