2.16.4 fixed a bug in memory allocation for band matrix
File network.cpp:698. The bug would not probably influence real time
gama-local input data and was discovered in random generated testing
data for sparse matrix algorithm `envelope`. In the case when in a
coordinate cluster the number of active unknowns was different from
the covariance matrix dimension, gama-local ended with a bad memory
allocation exception.