Fix the bug in test linearization visitor for Angle
The visitor did not test if the right hand direction of the angle
pointed to a fixed point or to an adjustment point. If it was a fixed
point visitor added corrections from vector x with index 0, i.e. index
outside the vector.

Added file bug/test-linearization-angle.gkf and .xml to test all
possible combinations of angles with fixed and adjusment points.

The bug was reported by Greg Troxel on netbsd 9 amd64.